+++ ggbk03/ggbk03a/bootpack.c
@@ -1,5 +1,4 @@
-#include <stdint.h>
-
+// #define directive
 #define COL8_000000 0
 #define COL8_FF0000 1
 #define COL8_00FF00 2
@@ -17,19 +16,18 @@
 #define COL8_008484 14
 #define COL8_848484 15
 
-extern void io_hlt(void);
-extern void io_cli(void);
-extern void io_out8(int port, int data);
-extern int io_load_eflags(void);
-extern void io_store_eflags(int eflags);
-extern void load_gdtr(int limit, int addr);
-extern void load_idtr(int limit, int addr);
+// From nasmfunc.asm
+void io_hlt(void);
+void io_cli(void);
+void io_out8(int port, int data);
+int io_load_eflags(void);
+void io_store_eflags(int eflags);
+void load_gdtr(int limit, int addr);
+void load_idtr(int limit, int addr);
 
-extern void mysprintf(char *str, char *fmt, ...);
-
+// Function prototype
+void mysprintf(char *str, char *fmt, ...);
 void init_palette(void);
-void set_palette(int start, int end, unsigned char *rgb);
-void boxfill8(unsigned char *vram, int xsize, unsigned char c, int x0, int y0, int x1, int y1);
 void init_screen8(unsigned char *vram, int x, int y);
 void putfont8(unsigned char *vram, int xsize, int x, int y, char c, char *font);
 void putfonts8_asc(unsigned char *vram, int xsize, int x, int y, char c, char *s);
@@ -81,205 +79,4 @@ void HariMain(void)
     for (;;) {
         io_hlt();
     }
-}
-
-void init_palette(void)
-{
-    unsigned char table_rgb[16*3] = {
-        0x00, 0x00, 0x00, //  0: 黒
-        0xff, 0x00, 0x00, //  1: 明るい赤
-        0x00, 0xff, 0x00, //  2: 明るい緑
-        0xff, 0xff, 0x00, //  3: 明るい黄色
-        0x00, 0x00, 0xff, //  4: 明るい青
-        0xff, 0x00, 0xff, //  5: 明るい紫
-        0x00, 0xff, 0xff, //  6: 明るい水色
-        0xff, 0xff, 0xff, //  7: 白
-        0xc6, 0xc6, 0xc6, //  8: 明るい灰色
-        0x84, 0x00, 0x00, //  9: 暗い赤
-        0x00, 0x84, 0x00, // 10: 暗い緑
-        0x84, 0x84, 0x00, // 11: 暗い黄色
-        0x00, 0x00, 0x84, // 12: 暗い青
-        0x84, 0x00, 0x84, // 13: 暗い紫
-        0x00, 0x84, 0x84, // 14: 暗い水色
-        0x84, 0x84, 0x84  // 15: 暗い灰色
-    };
-    set_palette(0, 15, table_rgb);
-    return;
-}
-
-void set_palette(int start, int end, unsigned char *rgb)
-{
-    int i, eflags;
-    eflags = io_load_eflags(); // 割り込み許可フラグの値を取得
-    io_cli(); // 許可フラグを0にして割り込み禁止にする
-
-    // [ビデオDAコンバータ](http://oswiki.osask.jp/?VGA#o2d4bfd3)
-    io_out8(0x03c8, start); // アドレスレジスタ (パレット書き込みモード)
-    for (i = start; i <= end; i++) {
-        io_out8(0x03c9, rgb[0] / 4); // データレジスタ
-        io_out8(0x03c9, rgb[1] / 4);
-        io_out8(0x03c9, rgb[2] / 4);
-        rgb += 3;
-    }
-    io_store_eflags(eflags); // 割り込み許可フラグを元に戻す
-    return;
-}
-
-void boxfill8(unsigned char *vram, int xsize, unsigned char c, int x0, int y0, int x1, int y1)
-{
-    int x, y;
-    for (y = y0; y <= y1; y++) {
-        for (x = x0; x <= x1; x++) {
-            vram[y * xsize + x] = c;
-        }
-    }
-    return;
-}
-
-void init_screen8(unsigned char *vram, int x, int y)
-{
-    boxfill8(vram, x, COL8_008484,  0,     0,      x -  1, y - 29);
-    boxfill8(vram, x, COL8_C6C6C6,  0,     y - 28, x -  1, y - 28);
-    boxfill8(vram, x, COL8_FFFFFF,  0,     y - 27, x -  1, y - 27);
-    boxfill8(vram, x, COL8_C6C6C6,  0,     y - 26, x -  1, y -  1);
-
-    boxfill8(vram, x, COL8_FFFFFF,  3,     y - 24, 59,     y - 24);
-    boxfill8(vram, x, COL8_FFFFFF,  2,     y - 24,  2,     y -  4);
-    boxfill8(vram, x, COL8_848484,  3,     y -  4, 59,     y -  4);
-    boxfill8(vram, x, COL8_848484, 59,     y - 23, 59,     y -  5);
-    boxfill8(vram, x, COL8_000000,  2,     y -  3, 59,     y -  3);
-    boxfill8(vram, x, COL8_000000, 60,     y - 24, 60,     y -  3);
-
-    boxfill8(vram, x, COL8_848484, x - 47, y - 24, x -  4, y - 24);
-    boxfill8(vram, x, COL8_848484, x - 47, y - 23, x - 47, y -  4);
-    boxfill8(vram, x, COL8_FFFFFF, x - 47, y -  3, x - 47, y -  3);
-    boxfill8(vram, x, COL8_FFFFFF, x -  3, y - 24, x -  3, y -  3);
-
-    return;
-}
-
-void putfont8(unsigned char *vram, int xsize, int x, int y, char c, char *font)
-{
-    int i;
-    unsigned char *p, d;
-
-    for (i = 0; i < 16; i++) {
-        p = (unsigned char *) (intptr_t) (vram + (y + i) * xsize + x);
-        d = font[i];
-        if ((d & 0x80) != 0) p[0] = c;
-        if ((d & 0x40) != 0) p[1] = c;
-        if ((d & 0x20) != 0) p[2] = c;
-        if ((d & 0x10) != 0) p[3] = c;
-        if ((d & 0x08) != 0) p[4] = c;
-        if ((d & 0x04) != 0) p[5] = c;
-        if ((d & 0x02) != 0) p[6] = c;
-        if ((d & 0x01) != 0) p[7] = c;
-    }
-
-    return;
-}
-
-void putfonts8_asc(unsigned char *vram, int xsize, int x, int y, char c, char *s) {
-    extern char hankaku[4096];
-    for (; *s != 0x00; s++) {
-        putfont8(vram, xsize, x, y, c, hankaku + *s * 16);
-        x += 8;
-    }
-    return;
-}
-
-void init_mouse_cursor8(char *mouse, char bc)
-{
-    static char cursor[16][16] = {
-        "**************..",
-        "*OOOOOOOOOOO*...",
-        "*OOOOOOOOOO*....",
-        "*OOOOOOOOO*.....",
-        "*OOOOOOOO*......",
-        "*OOOOOOO*.......",
-        "*OOOOOOO*.......",
-        "*OOOOOOOO*......",
-        "*OOOO**OOO*.....",
-        "*OOO*..*OOO*....",
-        "*OO*....*OOO*...",
-        "*O*......*OOO*..",
-        "**........*OOO*.",
-        "*..........*OOO*",
-        "............*OO*",
-        ".............***",
-    };
-    int x, y;
-
-    for (y = 0; y < 16; y++) {
-        for (x = 0; x <16; x++) {
-            if (cursor[y][x] == '*') {
-                mouse[y * 16 + x] = COL8_000000;
-            } else if (cursor[y][x] == 'O') {
-                mouse[y * 16 + x] = COL8_FFFFFF;
-            } else if (cursor[y][x] == '.') {
-                mouse[y * 16 + x] = bc;
-            }
-        }
-    }
-    return;
-}
-
-void putblock8_8(unsigned char *vram, int vxsize, int pxsize,
-    int pysize, int px0, int py0, char *buf, int bxsize)
-{
-    int x, y;
-    for (y = 0; y < pysize; y++) {
-        for (x = 0; x < pxsize; x++) {
-            vram[(py0+y) * vxsize + (px0+x)] = buf[y * bxsize + x];
-        }
-    }
-    return;
-}
-
-void init_gdtidt(void)
-{
-    struct SEGMENT_DESCRIPTOR *gdt = (struct SEGMENT_DESCRIPTOR *) 0x00270000;
-    struct GATE_DESCRIPTOR    *idt = (struct GATE_DESCRIPTOR    *) 0x0026f800;
-    int i;
-
-    // Init GDT
-    for (i = 0; i < 8192; i++) {
-        set_segmdesc(gdt + i, 0, 0, 0);
-    }
-    set_segmdesc(gdt + 1, 0xffffffff, 0x00000000, 0x4092);
-    set_segmdesc(gdt + 2, 0x0007ffff, 0x00280000, 0x409a); // for bootpack.hrb
-    load_gdtr(0xffff, 0x00270000);
-
-    // Init IDT
-    for (i = 0; i < 256; i++) {
-        set_gatedesc(idt + i, 0, 0, 0);
-    }
-    load_idtr(0x7ff, 0x0026f800);
-
-    return;
-}
-
-void set_segmdesc(struct SEGMENT_DESCRIPTOR *sd, unsigned int limit, int base, int ar)
-{
-    if (limit > 0xfffff) {
-        ar |= 0x8000; // G_bit = 1
-        limit /= 0x1000;
-    }
-    sd->limit_low    = limit & 0xffff;
-    sd->base_low     = base & 0xffff;
-    sd->base_mid     = (base >> 16) & 0xff;
-    sd->access_right = ar & 0xff;
-    sd->limit_high   = ((limit >> 16) & 0x0f) | ((ar >> 8) & 0xf0);
-    sd->base_high    = (base >> 24) & 0xff;
-    return;
-}
-
-void set_gatedesc(struct GATE_DESCRIPTOR *gd, int offset, int selector, int ar)
-{
-    gd->offset_low   = offset & 0xffff;
-    gd->selector     = selector;
-    gd->dw_count     = (ar >> 8) & 0xff;
-    gd->access_right = ar & 0xff;
-    gd->offset_high  = (offset >> 16) & 0xffff;
-    return;
 }
